feat: added search bar in departments table
