+++ src/components/companyDepartments/DepartmentEmployees.tsx
@@ -1,5 +1,4 @@
 import { Link } from "react-router-dom";
-import { Button } from "../ui/button";
 
 export default function DepartmentEmployees({ users }: { users: any }) {
   return (
