+++ src/components/companyDepartments/departmentsInfo.tsx
@@ -3,15 +3,17 @@ import { Separator } from "../ui/separator";
 import { SearchX } from "lucide-react";
 
 interface DepartmentsInfoProps {
-  companyInfo?: {
+  companyTotals?: {
     departments: number;
     employees: number;
     positions: number;
   } | null;
 }
 
-export default function DepartmentsInfo({ companyInfo }: DepartmentsInfoProps) {
-  if (companyInfo === undefined) {
+export default function DepartmentsInfo({
+  companyTotals,
+}: DepartmentsInfoProps) {
+  if (companyTotals === undefined) {
     return (
       <div className="h-36 w-full flex items-center justify-center border border-neutral-200 rounded-md">
         <span className="text-neutral-500 flex gap-3 items-center">
@@ -21,7 +23,7 @@ export default function DepartmentsInfo({ companyInfo }: DepartmentsInfoProps) {
     );
   }
 
-  if (companyInfo === null) {
+  if (companyTotals === null) {
     return (
       <div className="h-36 w-full flex items-center justify-center border border-neutral-200 rounded-md">
         <span className="text-neutral-500 flex flex-row gap-3 items-center">
@@ -33,24 +35,24 @@ export default function DepartmentsInfo({ companyInfo }: DepartmentsInfoProps) {
   return (
     <div className="h-36 w-full flex flex-col justify-center border  border-neutral-200 rounded-md ">
       <div className="flex-none  flex flex-row p-1 px-5 gap-5 h-full">
-        <div className="flex flex-col justify-center  flex-1 ">
+        <div className="flex flex-col justify-center   items-center flex-1 ">
           <span className="text-base font-light">Departamentos:</span>
-          <span className="font-black text-4xl text-neutral-700 self-end pr-5">
-            {companyInfo.departments}
+          <span className="font-black text-4xl text-neutral-700  ">
+            {companyTotals.departments}
           </span>
         </div>
         <Separator orientation="vertical" className="h-3/5 my-auto" />
-        <div className="flex flex-col justify-center  flex-1 ">
+        <div className="flex flex-col justify-center items-center   flex-1 ">
           <span className="text-base font-light">Cargos:</span>
-          <span className="font-black text-4xl text-neutral-700 self-end pr-5">
-            {companyInfo.positions}
+          <span className="font-black text-4xl text-neutral-700 ">
+            {companyTotals.positions}
           </span>
         </div>
         <Separator orientation="vertical" className="h-3/5 my-auto" />
-        <div className="flex flex-col justify-center  flex-1 ">
+        <div className="flex flex-col justify-center items-center   flex-1 ">
           <span className="text-base font-light">Empleados:</span>
-          <span className="font-black text-4xl text-neutral-700 self-end pr-5">
-            {companyInfo.employees}
+          <span className="font-black text-4xl text-neutral-700 ">
+            {companyTotals.employees}
           </span>
         </div>
       </div>
